add log to last print2log
