--- py/lightcone/jax_implementation/.gitignore
@@ -1,5 +1,0 @@
-__pycache__/
-validation
-memory_profile-bk.log
-mprofile_*
-
