+++ py/lptmap/.gitignore
@@ -1,0 +1,5 @@
+*.fits
+__pycache__
+displacement.py 
+test_slicing.py 
+testslicing.txt
