+++ py/lptmap/README.md
@@ -1,0 +1,10 @@
+**Needs `gputil` installed to use JAX with GPU**
+
+* To run the code: `python lightcone.py`. So `mpirun ... python lightcone.py` or `srun ... python lightcone.py`
+* Parameters are still hardcoded in `lightcone.py`.
+* Supports `LPT_DISPLACEMENTS_PATH` to pass displacement directory path
+* Full displacement cubes will need change on file names hardcoded in  `lightcone.py`.
+
+
+
+
